// src/types/generated/patientAppointmentsSummaryResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Doctaura API
 * API for managing clinic operations, doctor schedules, and appointments
 * OpenAPI spec version: v1
 */
import type { PatientAppointmentSummaryDto } from './patientAppointmentSummaryDto';
import type { AppointmentStatisticsDto } from './appointmentStatisticsDto';
import type { PrescriptionSummaryDto } from './prescriptionSummaryDto';

export interface PatientAppointmentsSummaryResponseDto {
  patientId?: string;
  /** @nullable */
  patientName?: string | null;
  nextAppointment?: PatientAppointmentSummaryDto;
  /** @nullable */
  recentAppointments?: PatientAppointmentSummaryDto[] | null;
  /** @nullable */
  upcomingAppointments?: PatientAppointmentSummaryDto[] | null;
  statistics?: AppointmentStatisticsDto;
  /** @nullable */
  availableActions?: string[] | null;
  /** @nullable */
  prescriptions?: PrescriptionSummaryDto[] | null;
}