// src/types/generated/patientAppointmentSummaryDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Doctaura API
 * API for managing clinic operations, doctor schedules, and appointments
 * OpenAPI spec version: v1
 */
import type { AppointmentType } from './appointmentType';
import type { AppointmentStatus } from './appointmentStatus';

export interface PatientAppointmentSummaryDto {
  id?: string;
  appointmentDateTime?: string;
  appointmentType?: AppointmentType;
  status?: AppointmentStatus;
  /** @nullable */
  statusDisplayName?: string | null;
  durationMinutes?: number;
  doctorId?: string;
  /** @nullable */
  doctorFullName?: string | null;
  /** @nullable */
  doctorPhotoUrl?: string | null;
  createdAt?: string;
  /** @nullable */
  cancelledAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}