// src/types/generated/prescriptionSummaryDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Doctaura API
 * API for managing clinic operations, doctor schedules, and appointments
 * OpenAPI spec version: v1
 */
import type { Medicine } from './medicine';

export interface PrescriptionSummaryDto {
  id?: string;
  appointmentId?: string;
  doctorId?: string;
  /** @nullable */
  doctorName?: string | null;
  patientId?: string;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  medicines?: Medicine[] | null;
  dateIssued?: string;
  dateCreated?: string;
  dateUpdated?: string;
}